updated ChatLazyColumn
